TST Move compareDictOfArrays to test/__init__.py
